requirements.txt
lxml
aliyun-python-sdk-core
aliyun-python-sdk-alimt
tqdm
colorama
PyYAML